[UNFINISHED] Make image load before applying any modifications
This was intended to lock the screen faster and update the image
whenever the modified version was ready.
All it did was give me a masive headache.
WHY DOES IT NOT WORK?
I mean it kind of does, but you have to update the window for it to show.
I'm new to all of this wayland and C stuff, so I have no idea what to do.